chore: refactoring in order to remove utils form test coverage
