+++ src/utils/testing/testRenderProviders.tsx
@@ -3,8 +3,8 @@ import { render } from "@testing-library/react";
 import type { RenderOptions } from "@testing-library/react";
 import { Provider } from "react-redux";
 
-import type { AppStore, RootState } from "../redux/store";
-import { setupStore } from "../redux/store";
+import type { AppStore, RootState } from "../../redux/store";
+import { setupStore } from "../../redux/store";
 
 // This type interface extends the default options for render from RTL, as well
 // as allows the user to specify other things such as initialState, store.
